LogApplier's persist function was calling callback with wrong args
When the LogApplier was done persisting it was calling the done()
function to signal completion, which is correct. However, it also
registered the callback directly to the 'done persisting' event which
was incorrect. The 'done persisting' event gets called with the last
applied commit index, which was being interpreted by the callback as
an error since callbacks take the form function(err, value).