Batch replication entries for greater throughput
This needs a configurable batch size - it's currently hardcoded to
16. To get replication batching working, the log applier code also had
to be modified so that we queue up calls to LA.persist and execute
them serially. Otherwise, the logic would see we're already in the
middle of a persist operation and not do anything, which means the
callback for that persist operation never got called.

So if we end up batching 5 commands together, we send one replication
command but trigger 5 persist callbacks. The first time persist gets
called it persists all the commands and the subsequent 4 times are
basically no-ops that just call their callbacks to keep the async
chain going.